Remove search batching.
It turns out that my approach to search batching was just fundamentally
bad! Because the UCT values were not updated, it didn't actually change
the values of the search. This meant that it was just evaluating the
same exact search 8 times -- essentially resulting in 8x fewer total
searches in 2x as much time.